trivia-front/back/pyproject.toml
[tool.poetry]
name = "trivia"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
flask-restful = "^0.3.10"
python-dotenv = "^1.0.0"
flask-sqlalchemy = "^3.0.5"
flask-migrate = "^4.0.4"
flask-cors = "^4.0.0"
openpyxl = "^3.1.2"
requests = "^2.31.0"
passlib = "^1.7.4"
flask-jwt-extended = "^4.5.2"
argon2-cffi = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
